refactor: update project configuration and CLI structure
- Changed project versioning to dynamic in pyproject.toml.
- Updated development status from Alpha to Pre-Alpha, since we are still in development mode.
- Introduced a unified Typer CLI with Rich logging for better user experience instead of using click
